requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
beautifulsoup4
requests
